factor out ParseAssertionsWithReaders
