+++ go/libkb/assertion.go
@@ -793,43 +793,3 @@ func ParseImplicitTeamTLFName(ctx AssertionContext, s string) (keybase1.Implicit
 	isPublic := parts[2] == "public"
 	return ParseImplicitTeamDisplayName(ctx, parts[3], isPublic)
 }
-
-// Unpack an assertion with one or more comma-separated parts. Only AND assertions are allowed within each part.
-func UnpackAssertionList(expr AssertionExpression) (res []AssertionExpression, err error) {
-	switch expr := expr.(type) {
-	case AssertionOr:
-		// List (or recursive tree) of comma-separated items.
-
-		if expr.symbol != "," {
-			// Don't allow "||". That would be confusing.
-			return res, fmt.Errorf("disallowed OR expression: '%v'", expr.symbol)
-		}
-		for _, sub := range expr.terms {
-			// Recurse because "a,b,c" could look like (OR a (OR b c))
-			sublist, err := UnpackAssertionList(sub)
-			if err != nil {
-				return res, err
-			}
-			res = append(res, sublist...)
-		}
-		return res, nil
-	default:
-		// Just one item.
-		err = checkAssertionListItem(expr)
-		return []AssertionExpression{expr}, err
-	}
-}
-
-// A single item in a comma-separated assertion list must not have any ORs in its subtree.
-func checkAssertionListItem(expr AssertionExpression) error {
-	if expr.HasOr() {
-		return fmt.Errorf("assertions with OR are not allowed here")
-	}
-	switch expr.(type) {
-	case AssertionOr:
-		// this should never happen
-		return fmt.Errorf("assertion parse fault: unexpected OR")
-	}
-	// Anything else is allowed.
-	return nil
-}
